Add patch by Laurent Bartholdi to help compute lower central quotients.
To quote from Laurents email, addressed to me and Eamonn O'Brien:

Returning to my question about how to compute lower central quotients:

It seemed the simplest thing to do was to modify anupq, as follows. In the
"tails" routine, compute for each consistent generator its "lower central
degree": the original generators get degree 1 (I'm assuming the group is given
without redundant generators), the p-power of a degree-n generator gets degree
n, and the commutator of degree-m and degree-n generators gets degree m+n.
Then when the code creates new tails, it creates all p-power tails of the last
class, and commutator-tails of all generators in the last class that have
degree < N. In this manner, it seems the modified anupq will compute
p-approximations to G/gamma_{N+1}.

I wanted to run this through you to make sure I didn't miss something stupid.
I don't know yet how to make the code cleanly process generators that happen
to belong to gamma_2.

Attached is the diff file wrt. anupq 3.1.5. The syntax is very simple: in the
advanced menu (8), when the command "add tails" (7) asks in which class tails
should be added, give a negative number, which will be treated as -N, rather
than "0" (all classes) or a positive number.

It seems to me that, in fact, this (if cleanly coded) would be a valuable
addition to anupq: bounding the lower central series class, in effect adding
the identity [X_0,X_1,...,X_c].

Another nice modification would be to add p-power tails of class-c consistent
generators in class pc rather than in class c+1. In this way, one would
compute quotients by the Zassenhaus-Jennings-Lazard-... filtration rather than
the exponent-p central one.