Require GMP and take GAP's GMP_PREFIX into account (#66)
There is no good reason to allow building anupq without GMP
these days. On the downside, if one fails to do that (which may
happen without the user even knowing what happened), one may get
unexpected errors or perhaps even wrong results. So better to
just require this.

This also simplifies the logic in configure.ac considerably,
and allows us to just use the same as in anunq.